📏 Ensure noise map outputs are always in the unit square
Previously I was overloading width and height to mean two different things. The width and height are properties of the underlying noise, and not related to the actual dimensions of the noise map. Since the noise map is in theory infinite resolution, the actual width and height make no sense.